Support weighted choices in the grammar
Adds syntax node construction and evaluator for weighted choices. This
is much the same functionality as the Ruby version, but more restricted
in scope:

- Ranges as weights are not supported as JavaScript doesn’t have syntax
for ranges built-in
- Two element arrays are unsupported as this syntax is confusing and
unncessary—the favoured approach (which is probably what will become the
spec) is to use key/value pairs via hashes/objects

Symbol table sharing via the registry class is not yet supported, so
there’s currently a hack to make the RNG injectable.