Add expression parsing to concat nodes
Modifier chains are now supported in expressions. Syntax nodes delegate
to the registry to run the actual filters, which means it will be
straightforward to add injectable/customizable modifiers soon.

For now, only basic JavaScript String methods are supported
(toUpperCase, toLowerCase). The parser also makes room for memo and
unique nodes, which are not yet implemented.