Adding memoization sigil to expression parser
Memoized production rules are now supported, in the exact same way as
the Ruby version. Note the method rename in the port, from `expandMemo`
to `evaluateMemo` as this reflects the semantics properly. The original
method was misnamed—in order to return the same subtree result each
time, the entire rule must be evaluated the first time it’s
selected—otherwise the top level branch will be the same, but lower
level branches will differ each run. This would be confusing.