Add unique cycle sigil to production parser and registry
The syntax tree can now expand unique cycle references in production
rules. This is a little bit awkward because of the brute force loop
ported from the original Ruby implementation. There are probably more
elegant ways of doing this, but would involve some larger changes to the
architecture. Something to consider for version 2.

There’s also some potentially problematic behaviour with expansions of
deeply nested rules using the @ and $ syntax. Most of the assumptions
and tests written cover the basic case where the rule terminates and it
makes a lot of sense for one level of choices. Where choices are deeply
nested and there are more permutations to consider, this may lead to
problems. This is where a more thorough language specification would
come in handy.