Ensure memo and unique nonterminals get correctly compiled to AST
Fixes a bug in the concat parser where the full syntactic string was
being incorrectly assigned as the nonterminal reference, rather than a
qualified rule definition. Probably a typo or copy/paste accident.

Main thing here is this highlighted the lack of high level tests of this
syntax from the outside. Individual unit tests of the internal parts
didn’t catch this.