Upgrade to Ava V3 and add full paths to imports
Now that native ESM is supported everywhere, it’s inane and pointless to
pursue the React-era obsession with extensionless file imports. If
adding a simple .js at the end of each import is all that it takes to
get this entire thing working properly in Node and browsers without
needing to transpile a whole bunch of shit, then this is very good.