Refactor codebase to pass options down the evaluate tree
This is a feature of the codebases in the other languages. Was never
done in JS because having an Options class/type felt heavyweight and
pushing against the grain. But probably not a big deal. You can write OO
JS just fine, but the style here is very 2018 function wrappers for
everything. So a bit more thought is required to clean up the internals.
For now, the outside API doesn’t change but it should be possible to
seed the RNG if necessary.