Scope memo expansions to sub-branches rather than globally
This is an experimental new feature to enable memo expansions to be
scoped to branches, meaning that the memos can be set multiple times for
a single grammar run and will remain consistent throughout subtree
expansions, but reset at higher levels of the tree.

```
{
  start: "{report};{report}",
  report: "{@@name}:{description}",
  description: "{@@name}",
  name: ["fluorite", "pyrite", "sodalite"]
}
```

There are some problems with this implementation so it can’t be merged
as-is, but it’s hopefully a useful starting point for further discussion
both around the context of the feature and syntax change, and about the
implementation.

What I’d also like to consider is if some of this functionality can be
moved from the registry into the syntax nodes so that some of the random
selection details can be better encapsulated, rather than letting the
registry drive nearly all the behaviour.